Create a Retire model
This will be used to stored details of why a document was retired

Created by running:
`rails generate model Retire explanatory_note:string timeline_entries:references`

Retire is a child of TimelineEntry. As the plural of TimelineEntry
doesn’t follow the rails convention of just appending an `s` we also
have to specify the foreign key id.