Set hard max lengths on title and summary inputs
This sets a maxlength on the textarea to be the maximum we accept from a
user. This is to stop users entering a huge amount of text when we'll
only reject it at the server. Users shouldn't really be hitting this
under any normal circumstance since they are warned from far lower
numbers to enter less.