Handle topics 404ing for a document
If we haven't wrote a document to the Publishing API then any call to
get the links for it will return a 404. This was previously getting
picked up in an exception handler in the view, but since this is a
pretty normal part of a documents flow then it makes sense to be handled
in the DocumentsTopic itself.