Create a Retirement model
This will be used to stored details of why a document was retired

Created by running:
`rails generate model Retirement explanatory_note:string timeline_entries:references`

Retirement is a child of TimelineEntry. As the plural of TimelineEntry
doesn’t follow the rails convention of just appending an `s` we also
have to specify the foreign key id.