Replace vendored and polyfill.io with packages
Previously we used a mixture of vendored JavaScript and polyfill.io to
import external JavaScript. polyfill.io doesn't always detect the right
polyfills to apply and isn't available on certain internal networks.

This increases the use of package.json to install the necessary
poltfills from NPM and removes the previously vendored files and
external call to polyfill.io.