Use core-js to fix topics and replace polyfills
Previously we removed the use of Reflect from the markdown toolbar and
its associated, broken polyfill. Unfortunately the miller-columns
component also makes use of Reflect. As the original polyfill is broken,
the only alternative package seems to be core-js(-bundle), which also
makes a few of the other polyfills redundant (e.g. promise-polyfill).