Add download link to images
Clicking the download link will crop the image to 960x640 and download
it to the users computer. This is useful for whitehall users as images
uploaded here can easily be cropped and used in whitehall publisher.

I've added a bang to the end of the crop dimensions as this enforces
the images are cropped to the exact size specified (certain images can
be a few pixels out due to minimagick trying to retain the exact aspect
ratio). -> http://www.imagemagick.org/Usage/resize/#noaspect

Due to a discussion with Sonia, the link order has also been amended
which moves the `delete image` link to the end of the list.

Trello:
https://trello.com/c/w4aI961C/529-provide-a-960x640-download-link-for-images