Move warn-before-unload to new JS modules directory
This javascript isn't an example of a GOV.UK component as it is
something that can be applied to components rather than a component in
it's own right. In terms of components this would belong as an option of a form
component, but I think even then you'd want this code separate so it
wasn't coupled to a form component.

In light of this, after a chat with Alex,  I've taken the bold step of creating
a modules directory to put things that can apply to elements and components
that aren't actually components.