Refactor warn before unload module
This makes a few changes to the warn before unload module:

- It now is configured to loop through forms and set up module on each
individual form - for greater consistency with component initialisation
- It uses a global variable on the event handler so that this function
can't be set/removed multiple times
- It sets a message so that any legacy browsers that still support
messages here receive one

I've also added a basic test for this module. It's not a super easy
thing to test since you don't actually want an alert to happen in your
tests and you can't check the message, but hopefully it's sufficient.