Push submited radio inputs into GA's dataLayer
If a form is given the data-attribute gtm then GTMFormListener will
listen for the form to be submitted and push an event to the
dataLayer for each checked radio input. Events are in the form

  { gtm-form-attribute: { radio-name: radio-value } }

This tracking is added to the new document forms and also the
confirm that a document has been reviewed form.