Replace attr_accessors by inerhiting ReadonlyModel
We have a number of classes, such as DocumentType and Topic, that
act as interfaces to static (readonly) data. Previously, some of these
classes used attr_accessors so they could include ActiveModel::Model,
which helps to DRY-up the way their instances are constructed.

This replaces the use of attr_accessor with attr_reader by changing each
model to extend a slightly modified version of ActiveModel::Model that
also makes the update_attributes and attributes= methods private.