Add current and live flags to an edition
This is a boolean field that is set on an edition that expresses whether
an edition is currently the live or current version of a document.

The current edition of a document is the edition that a user should be browsing
when interacting with Content Publisher, most likely this is the latest
version of the document - although this might not be the case if the
user has discard the latest draft.

The live edition of a document is the edition that is currently on
GOV.UK in some form, it may also be the current item.