Create basic revision table
This would store each of the revisions of an Edition. It stores quite a
lot of data in it so there could be a bunch of duplication between
records but it is simpler than creating more models.

Initially I was going to store an association to an edition here,
however we have a chicken and the egg situation with a edition and
revision association where an edition needs a current revision and a
revision should know about it's edition. I decided that since it could
feasibly be possible that a revision is associated with two editions
and that we never want an edition not to have a current edition it
belongs best on an edition.