Create a many to many join between revision and editions
An edition will always have a current revision, but it's less clear that
revisions always have just a single edition (and also we can't write
both at the same time)

Thus I thought the cleanest way to resolve the chicken and egg problem
there was to have a many to many association. This therefore allows the
possibility that a revision could be associated with multiple editions
(which would make sense say when you first create an edition or if you
wanted to roll an edition back to a revision). This association can also
be created after edition and revision exist avoiding the chicken and egg
situation.