Handle Publishing API being down when withdrawing a document
This ensures that an error message is displayed to users in the event that they
attempt to withdraw a document but the Publishing API is unavailable and unable
to process the request.

The Unpublish Service wraps withdrawing a document in a transaction so if the
Publishing API is unavailable then the document's state is unchanged.