Add entry_type of `internal_note` for timeline_entries
This adds the ability to create a timeline_entry with a type of
`internal_note` and display the note on the document history page.

Frontend changes:
I've also swapped the timestamp to come before the internal note /
explanatory note as this groups the entries a little clearer.

I've added a two thirds column to prevent long sentences as per the
prototype design, this also leaves space to add the internal notes
button.

The body of a internal note is rendered with links intact. This is done
via `simple_format` which converts `\n` to `<br>` and the gem rinku
which identifies URL's and email address's and creates hyperlinks and mailto
links respectively. All html tags are also escaped to prevent users
adding their own html / browsers mistakenly thinking < > etc are html tags.