Refactor image upload requirements as a checker
This refactors the code to compute issues for image uploads so that it
matches the checker approach we use everywhere else. Part of this
approach is that we only check for one or two issues in the feature
specs, and use a unit spec to cover all of them in one place.