Separate document type intialize from deserialize
Previously the constructor for the DocumentType model was also
responsible for deserializing the data from its YAML hash form. This
separates the deserialize and initialize concerns by adding a 'from_yaml'
method on DocumentType, which deals with any postprocessing.

Separating the deserialize and intialize concerns means we can now make
the factories for document type and its inner classes consistent in
constructing instances of the models they describe, rather than hashes.

foo