Correct the factory for document types
Previously we used the initialize_with part of the document type factory
to do some post-build tasks and also used the 'add' method of document
type to construct an instance; having an 'add' method on document types
also unnecessarily suggests this functionality is used in the app code.

This moves the behaviour of 'add' into the document type factory and
splits the code into initialize and post-build concerns.