Fix secret preview URL spec for DocumentUrl
The new version of JWT bundled with gds-sso refactors the way the
encoded string is computed. Part of the encoding involves hashing the
'headers' argument, which used to be defaulted to

     { "typ" => "JWT", "alg" => algorithm }

but is now just { "alg" => algorithm }. I don't think this impacts the
secrecy we're aiming for, so it should be OK to just update the spec.