Add basic unwithdraw workflow
We need to show the option to undo a withdrawal to users when a document has
been withdrawn, however the functionality for it has not been implemented yet.
This adds a simple path so that if users click to undo a withdrawl they
presented with the guidance to raise a support request.