Store a reference to document on a revision
Previously revisions had an implicit association with a document through
editions. This creates a concrete one.

The reason for adding this is to make the concept of a revision clearer.
Every edit of content in a document is a new revisions and the history
of document changes can be traced through revisions. Editions are used
to present a particular revision to the Publishing API.