Display withdrawn data in summary metadata
When a document has been withdrawn we display only the withdrawn state, last
updated date and name of the user who withdrew the document in the metadata
section on the document summary page.

In order to be able to display the withdrawn by user we ensure the current user
is saved to the Timeline Entry when the document is withdrawn. This required a
slight change to the Unpublish Service, so this commit also contains an update
to the unpublish rake task in order to comply with this change.