Create base withdraw document page
This creates a Withdraw controller and associated view and translation.  At the
beginning of building this app we had decided to use 'Retire' however it has
now been decided that the existing Whitehall terminology of 'Withdraw' should
be used instead.  For now we are keeping 'Withdraw' behind a pre release
feature flag, so we continue to use the legacy retire route until the feature
flag is removed.