Update last_edited_at/last_edited_by storage
This adds in last_edited_by for documents and editions to track which
user performed the action.

It also updated last_edited_at to no longer support nil and sets a value
for it as part of a before_create callback.

Initially I tried not to use the before_create callback as I hoped I
could achieve the goal with a database timestamp.

I tried the new feature in Rails 5 of setting:

```
change_column :versioned_editions,
              :last_edited_at,
              :datetime,
              null: false,
              default: -> { "CURRENT_TIMESTAMP" }
```

However when using this you'd need to reload an object to get the value
eg:

```
m = model.create
m.last_edited_at
=> nil
m.reload
m.last_edited_at
=> Fri, 21 Dec 2018 23:44:33 GMT +00:00
```