Create EditionStatus model
This creates a model to store the status of an Edition. For each
user_facing_state an edition is in there should always be a
EditionStatus for it.

On an EditionStatus we store:

- user_facing_state - which is what a user would see
- publishing_api_sync -  which is used to store whether the
  status has synced with Publishing API
- revision_at_creation_id - the revision of the edition that was current
  at the point the state was created (so we know which revision was
  submitted_for_review for instance)
- created_by_id - the user which changed the status
- edition_id - which edition the status is associated with, so we can
  look up all statuses an edition has had.

On an Edition we store:

- status_id - which is the current status that Edition has.

We hit something of a chicken and egg situation with regards to storing
edition on status and current status on edition. An edition should
always have a status so this is a non-null field, because of this a
status must be able to exist before an edition is created, therefore the
edition_id on a status is nullable so this can happen. For ease this
value is set automatically in an ActiveRecord callback when Edition is
saved.

The values for user_facing_state and publishing_api_sync are both
defined as enums. The reason they are set with strings rather than the
Rails default of integers is so that they are still understandable
without the ActiveRecord abstraction layer, which reflects concerns
raised by GOV.UK developers about the integer usage.