Remove publishing_api_sync field for edition draft field
This removes the publishing_api_sync field that was on the EditionStatus
field, as it had a number of problems:

- It makes no sense on a past status
- An edition could be synced to Publishing API draft but not live
- A change to an edition changes this status each time

I think the main flaw in it was that it hoped to do two rather distinct
things: 1) track whether a status change was successful and 2) track
draft syncing status.

These actually are rather different as while a document is in draft we
don't care about the Publishing API for any state changes - e.g. if
you've submitted a document for review it doesn't really matter if
Publishing API is up to date - and it may be synced multiple times in
that process.

Whereas if you are publishing or removing a document it doesn't make
sense to have that state until that operation is complete, and it also
seems natural that we'd want to store that someone tried to publish a
document and failed.

The approach I've looked at for this is to create a `draft` enum field
which is only applicable when a document is in one of the pre-publish
states. When it is one of those states it can have a value of
"requirements_not_met", "available" or "failure", when not in those
states it can have the "not_applicable" value.