Versioned document actions and metadata
This updates these view partials to work with the versioned documents
model.

In actions it drops support for a couple of states:
error_discarding_draft and error_sending_to_live since these can
hopefully be just shown if the operation fails as a flash rather than
being an aspect of state.