Versioned discard draft approach
This creates the approach for discarding a draft when we have documents
with editions. It creates a new state called discarded to be given to an
edition that has been discarded.

When an edition is discarded it is no longer the current edition of a
document, thus if there is a live edition it becomes the current edition
otherwise the document has no current edition.