Change Image model to ImageRevision model
This renames the Image model to be ImageRevision to reflect that this
refers to a revision of an image rather than a continuous item. It
creates a Image model to act as a continuous item - without this only
the association to a blob was a way of knowing that two image revisions
were related.

This change allows us to use the image_id as a user facing value so that
the id of the revision can be concealed. E.g. if we have image_id of 1
and image_revision_id of 2. This could have a URL of /images/1 for
image_id or /images/2 for image_revision_id. If we then created a new
revision /images/2 would no longer being a URL that works, whereas if we
use /images/1 that can work no matter what revision.