Many-to-many image_revision asset_manager_files relationship
This creates two new models: AssetManagerFile and
AssetManagerImageVariant. Where AssetManagerFile is intended to
represent the state of a file on asset manager and
AssetManagerImageVariant is used as a many to many join between an
ImageRevision and AssetManagerFile.

These changes introduce the concept of asset manager variants existing
for a single ImageRevision. This is to reflect the fact that we actually
have to upload multiple files in different sizes for a single image
which was previously de-prioritised or overlooked in this codebase.

AssetManagerImageVariant stores 3 fields:

- a reference to an ImageRevision
- a string to describe which variant it is (these are defined as 300,
  960 and hi_res to reference different image sizes)
- a reference to an AssetManagerFile

A unique index ensures that the same variant can't exist for the same
image revision.

AssetManagerFile stores:

- the URL of the file on asset manager
- the state in asset manager: absent, draft or live

A unique index ensures we can't have a file with the same URL stored
twice.

Whenever we create a new ImageRevision we also create
AssetManagerImageVariant for each variant of it. The associations to
AsseAssetManagerFile are either maintained or newly created depending on
whether the change to an Image has altered the image or not (if a user
changes the crop we need to re-upload it, whereas alt-text needs no
changes)