Port AssetManagerService for versioning
This changes the input to the AssetManagerService to have a method of
bytes on it and have auth_bypass_id as a parameter rather than taken
from a document.

This allows us to pass in an AssetManagerImageRevision rather than the
Image of before, which means we don't need to pass in bytes as an
argument.