Port preview controller for versioning
The preview creation is now in a transaction as creating the previews
will modify the document edition. I tried to maintain the rescue for the
whole method which is actually written a little bit dubiously as it
relies on the document variable being initialised. To counter this I've
set it to be an instance variable, which doesn't seem ideal but a bit
nicer than adding a begin / end