Add superseded concept to assets
When we have new versions of assets uploaded it's prudent to redirect
the old asset to the new one. To represent this there is a state of
superseded for an AssetManagerFile and an association to track which
file was used to supersede it.

I don't expect it'll be used for much more than debugging so I've not
indexed it.