PublishService for Versioned Documents
This is quite a lot different compared to the non versioned one since
we're dealing with both current and live editions of a document and
their assets.

The assets are quite complex so I switched to dealing with images rather
a more generic asset. These now behave by only making the lead image
live and removing the images of the previous document.

For a case where we have a newer version of an image we set the old one
to redirect at the point of publish so any usages of the asset don't
break.