Create Versioned::TimelineEntry
This model is used to represent TimelineEntry in a Versioned
ContentPublisher. Of all the models created for Versioning this is one
that seems most likely in need of iteration and shaping. At the time of
putting this together the needs of the timeline were somewhat in
flux/under negotiation so this tries to be mostly compatible with the
previous one but have additional associations to objects that
represent state at a particular time.

There are two associations that at this moment in time are speculative.
The association to revision and to status:

- The association for a revision is to be used when an edition content changes
  and would be expected to be used for diffing. It is also there so that
  if we decide we want to make the timeline entry notes different the
  concrete change can be determined to explain this. This association is
  not to be used for status changes.
- The association to a status is for the timeline entries that refer to
  a status change. A number of timeline entries refer to specific ones
  of these (e.g. published) sometimes these have additional context such
  as "Approved" to refer to a particular transition. This association is
  to allow these to have access to the definitive status information.

The reason they are here despite being speculative is that adding them
now in case is significantly less painful than trying to add them in
retrospectively if we determine we need them. To reflect the low usage
expectation they don't have indexes.

The entry_type of create_preview was removed as I don't think that needs
to be something that is shown on the history page and two new entries
have been added:

- draft_discarded: for when a live documents draft is discarded - this
  was previously not possible on Content Publisher
- draft_reset: for when an Edition is resumed after a draft has been
  discarded.

I pondered a bit about whether draft discarding and draft reset should
be part of the history. We had a history entry when a user creates an
edition so if that is discarded we have to decide whether to have a new
history entry to say that has been discarded or to remove chunks from
the history. I decided rather than deleting history it's more
transparent to keep the history there (and less confusing if someone
else discards your draft unexpectedly) however this probably cause some
minor WTFs when a user clicks "Create new edition" on a document that
has had a discarded draft.