Store and show timeline entries in versioned controllers
This serves as probably the fastest way to put these together. I divided
the types of timeline entry by whether it was a state change or a
revision. There were some cases where it acts as both as these (such as
creating a new edition resets change note) and with those I went with
status change as the record stored.