Create models for storing specific revision data
This creates three new models: ContentRevision, UpdateRevision and
TagsRevision which are associated with the Revision model. This then
removes a number of fields from Revision (title, summary, base_path,
contents, tags, update_type, change_note)

The intention of this is to create a storage efficient way to store all
the revisions a document goes through by storing the actual data in
specific models and then using Revision as a collection of joins. Thus
every time a user changes an image a new revision is created but that
can maintain a relationship to the ContentRevision so the data for that
does not need to be stored multiple times.

A user can interact with Revision in much the same way. The methods,
title, summary, etc are all delegated to their specific revision so you
can use `revision.title` to get that information without worrying about
the content revision. New revisions can be created with
`revision.build_revision_update({ title: "Test" }, user)` and it'll only
automatically build a new ContentRevision.

This also changes the controllers that mutate the state of a document to
check whether there has actually been a change to the data before
persisting a new revision. This has been added since, with us creating
revisions, it incurs storage each time a new revision is created so it'd
be good if they stored something different. This also means that for
every revision there is a difference that can be determined.