Create removal / retirement versioned models
This ports over the removal / retirement models for use in a versioned
context. These are associated with an EditionStatus models by way of a
polymorphic association.

These models are immutable and meant to effectively just be metadata to
an EditionStatus. In theory we could get rid of these being models and
instead use a JSON field, I'm not sure what is the path of least
surprise here - although polymorphic associations feels more Rails like.

Unlike the non versioned iterations of these models these ones don't
have associations to TimelineEntry directly and are instead associated
with that through the details polymorphic association on a
TimelineEntry.

These models are both meant to represent a status at a particular point
in time and are thus immutable. Were we to need to update the
information in them new instances should be created with a new status to
associate with.