Port rake task to remove/retire for versioning
This ports the task over with a few changes.

The biggest and most complex of these is a decision point about how to
treat assets based on whether a document has a draft edition. When a
document has a draft and a live edition they can share the same images.
When we remove an edition we plan to remove the images, but if the draft
also uses the images that can break the draft, thus we check which
images are used and decide whether to remove or draft them.

Otherwise, there's some shortening of the names of rake tasks and some
attempts made to make the tests a bit less verbose.