Port internal notes model to versioning
This is now associated with an edition rather than a document as we
don't really support timeline entries to not have an edition and since
these are displayed amongst them it seems to make sense to associate to
current edition.

This decision could all be reversed if the designs have internal notes
distinct from editions.