Rename EditionStatus to Status
The name of EditionStatus felt out of place as we don't have other parts
of an edition prefixed. I couldn't think of a better prefix and felt
that status was justified in it's grandness since this is the status of
this app cares most about. Hopefully if there is a competing term for
status in this app then we can think of an appropriate prefix (or at
least in that case EditionStatus won't feel strange)

This also renames the field user_facing_state in a Status to be just
state, this is because we don't have the same concepts of review_state,
live_state, publication_state in this versioned model and so this also
felt a bit strange in how it stuck out. I did initially want to call it
"value" to avoid the confusion of "Status" and "state" but sadly that's
a reserved word for an enum. Again, if this proves to be actually more
of a pain than user_facing_state we may have to change it back but I
think it should make sense for the most part.

This also adds the state enum methods to an Edition so now we can do:

edition.draft? rather than edition.status.draft?

Anyway fingers crossed this is a net benefit to the app...