Update edition status foreign key delete
This updates the on_delete cascades deletes to allow edition to be
deleted. This is to fix an issue where it's very difficult to
delete any data because of cyclic foreign key dependencies.

Although we don't anticipate the application to need data to be deleted
under normal use, this is done to not box users in for doing so in
abnormal usage.

With this change edition records are free to be deleted. Revisions of
that document can then be deleted and once they are gone the document
itself can be removed.