Fix Versioned::Document#newly_created?
Since removing the last_edited_at field from Document it is no longer
frequently updated so can't be used to check whether it is newly
created. This instead uses current_edition as the means for this check.

Currently this method is only used to make editing a document look like
you are creating it, hopefully we can amend this so we actually have a
create step and no longer need this.