Set inverse association of revision_at_creation
This sets a collection called statuses for a revision which tracks which
statuses were created when the document was at this revision.

The expectation of using this is in presenting debugging information to
show what statuses were created for a revision.

I wasn't sure if statuses was the best name since it doesn't capture the
"at_creation" part - but since I couldn't think of a better name I just
hope that doesn't cause too much confusion.