Re-instate revision number
This field was initially in the proposals for the versioning database
schema but was then dropped due to the expectation of not needing it.

When trying to think about how to set up a debug endpoint it became
rather apparent it'd be a useful number to have so it's returned. This
is based somewhat on the expectation that'd be way more painful to add
later than it is to have from the start.

This number isn't anticipated to be available to end users initially.