Fix bug with comparing image revisions
This used a sort method to check two arrays contain the same items. This
doesn't work if there are any null values. As we only expect the current
subject to be new we don't check that the other_image_revision_ids
contain null, as if that was to contain new records this method of
comparison is inappropriate.