Associate revisions with statuses
This replaces the many to many association between revisions and
editions with one between revisions and statuses. This association is used for
auditing processes and does not reflect the current state of the application.

The reason for this was that it was found that you couldn't determine
the statuses a revision held and that if you were to store this data you
could also work out the edition as status maintains a list of editions
stored.

Default orderings have been applied to these new associations to produce
consistent outputs.