Create a versioned debug page
This is a rather hacky debug page that is used to show the difference
between each of the revisions of a document.

It provides information on what data changed, which editions the content
is for and status' held by that revision.

To work out the difference it uses the HashDiff gem.