Fix image revisions blocking revision deletion
Prior to this change you would be unable to delete entries out of the
revision table if there were any associations with image revisions. This
might make these prohibitively hard to delete, instead we decide since
revision is the more important concept that if you want to delete that
then lose the joins.

We don't however allow the ability to remove an image revision without
removing the join first, this is because image_revisions are part of the
state of a revision and these should only be removed (without removing
the revision) with caution and exceptional circumstance.

Like many of these on_delete behaviours these are not here for day to
day to usage of the application, and are mostly anticipated to get used
in debugging or development scenarios.