Raise errors in UnpublishService when drafts are used
When this was originally wrote there was an expectation that you could
use it to remove/retire a live version while there is a draft present.
It transpires that this is actually not possible in the Publishing API,
the options you have there is either you can transition a draft to
unpublished (with the allow_draft option) or you can discard the draft
at the same time as unpublishing the live edition (with the
discard_draft option). Without one of these Publishing API will error
422 with:

"error":{"code":422,"message":"Cannot unpublish with a draft present","fields":{}}}

Thus this changes this code here to do checks on whether we're actually
trying to remove a live one and that a draft is not present.

As part of this work we can remove the rather complex work to check if
an asset should be removed or drafted depending on it's usage, and we
can remove the draft method on AssetManagerService as it is no longer
used.