Change database modelling for image revisions
After a long session with Ben Thorner chatting about concern/confusion
with the way we model Images with Image, ImageRevision, AssetManagerFile
and AssetManagerImageVariant, we came up with a different approach that
was more consistent with the modelling of other revisions in Content
Publisher.

In this we have:

Image - stores a couple of simple bits of continuous information such as
an id, who created it and when.

ImageRevision - stores references to image, file_revision and
metadata_revision, doesn't really too much - is just used to group data
together and be a common join point

ImageFileRevision - this stores data that changing means that the assets
created need to be changed. It stores filename and crop information.
Originally this was to be called CropRevision however it needed to be
expanded to include filename and various blob details.

ImageMetadataRevision - this stores data that can be updated without
needing to re-send the assets to Asset Manager, it stores information
such as alt text, credit and caption

ImageAsset - this has a many to one relationship with an
ImageFileRevision, there will be one of these for every resulting file
that is to be sent to asset manager. It stores variant which is a string
to indicate which type of file is sent, the state on asset manager and
the file_url from asset_manager. It also permits storing a superseded_by
reference for when a different image_asset replaces it. Finally there
are two unique indexes, one on file_url so we don't have two images
pointing to the same asset manager file and variant
image_file_revision_id compound so we can't store the same variant twice
for a file_revision.

Note: a previous version of this commit attempted to store non changing
image data such as blob, width and height on the image object however
this proved very problematic for ImageAsset which is associated with a
ImageFileRevision. Since ImageRevision knows about ImageFileRevision but
not vice versa this makes it challenging to know anything about the
underlying image.