Refactor the Versioned::Image modelling
This implements the models involved in image uploading as per the
previous commits database changes.

This changes ImageRevision to be split into three models: Image::Revision,
Image::FileRevision and Image::MetadataRevision. Image::Revision is now
used as a rather basic model that is used as a join point, in a similar
vein to Revision for a Document. The two separate models FileRevision
and MetadataRevision are split into data that changes the resultant file
(file revision) and data that doesn't (metadata revision).

By doing this it allows us to get rid of the many-to-many join we had
with AssetManagerFile through AssetManagerImageVariant. We don't need a
many-to-many anymore as every instance of Image::FileRevision has
distinct files on asset manager. This is represented through a one to
many association with Image::Asset which replaces AssetManagerFile and
AssetManagerImageVariant.

A pain point discovered in implementing this was that an
Image::FileRevision doesn't have knowledge of it's Image::Revision and
thus Image (this information flows the other way), this meant a need of
all file data to be on a FileRevision such as blob, width and height.
This doesn't seem to cause any problems - but does mean all data needed
for asset manager needs to be available in an association between
Image::FileRevision and Image::Asset. Maybe one day this will sting? Or
maybe this is actually a nice encapsulation. The alternative to this
route would be to go back to a many-to-many with an Image::Revision.