Replace draft column on versioned_editions
This replaces the string column on versioned_editions with a
revision_synced column. This column is to be of most use in a
draft context where we allow state changes and edits while a document is
not on the Publishing API. Whereas once we are in the realm of live
states the Publishing API must be synced for the state to be valid.