Replace usage of the Versioned::Edition draft enum
This changes using the draft on an Edition to use the revision_synced value.

By losing the enum there are less states that can be captured, this
means that we can no longer distinguish between a draft not being sent
to Publishing API because of lack of requirements and one that failed.
This means that we can no longer show the same "We failed to create a
preview message" that persists in the actions menu - but this doesn't
seem too great a concern as that's a bit of a strange thing to persist.