Refactor image_list partial to render just one
This moves some of the HTML out of the image_list partial into the
images index page, which is the only place it's used. This makes the
image_list partial simpler, to the extent that it's actually just a
partial for rendering a non-lead image on the image index page.