Rename UpdateRevision to MetadataRevision
This hopefully provides a more useful name to follow for the model which
stores the information such as update type and change log. The previous
name UpdateRevision was not remotely self explanatory and although I'm
not sure MetadataRevision tells a whole lot more it does seem an
improvement.

It is also consistent in naming with the Image::MetadataRevision model.

To follow on from this are database migrations that change the
underlying table and column names. As we have a batch of these to do and
they may momentarily cause application errors I intend to group these.