Turn off the Rails/HasManyOrHasOneDependent cop
This cop enforces specifying dependent in all has_one/has_many
relations. This is not always the right thing to define though as some
of these will be handled automatically by the database constraints.

For example if we have a `on_delete: :restrict` database constraint then
there's not much value in having a `dependent: :restrict_with_exception`
as the user will receive an exception from the database anyway - however
only the database exception is one we can guarantee happens
consistently.

Similarly for `dependent: :delete_all` is a dubious choice to use as it
won't call callbacks on any of the items we delete. We could go down a
road of putting `dependent: :destroy` which would mean all callbacks are
run each time, but even that feels like a presumptious and risky way to
manage database consistency.

Therefore this callback won't be used for maintaining database
consistency unless we are able to identify future needs for it.