Delete old tables superseded by versioning
This removes the tables that have been made redundant by the new
versioning code, and their associated foreign keys. We will rename the
new tables to omit the 'versioned' namespace in a subsequent PR, noting
that this change will cause some downtime.

The migration is deliberately non-reversible because these tables are
unused, so we do not expect to need to rollback for any reason.