Make feature tests more edition centric
This commit aims to bring more consistency to feature tests and
hopefully improve the comprehension of them.

Firstly, it makes consistent that we create an edition at the start of a
test. This replaces the inconsistent usage of some tests having
`create(:document, :with_current_edition)` with `create(:edition)`.

Secondly, it updates the methods in tests to be more edition specific
replacing nearly all `given_there_is_a_document` with
`given_there_is_an_edition` and similar. This tends to actually make
more sense as most of the operations we do are actually on an edition.
E.g. "When I publish a document" is actually less exact than "When I
publish an edition".

Finally, it reduces the usage of the word document around the tests,
most notably the many references to the "document page" which has now
been replaced with "summary page". This was mostly done to avoid
confusion with the word document sprouting out, but overall it seems to
be an improvement anyway as document page is rather ambiguous.

This only does what it can in feature tests to resolve the document
centric nature of wording though, as the app is mostly phrased around a
world before editions. So various bits of document terminology is still
present.