Move the unpublish rake test
This moves the test from the spec/features/workflow directory into the
spec/tasks directory. This test was inconsistent with our other feature
tests.

It wasn't obvious what the place would be for it as it something of an
integration/feature test but as we have strictly defined the spec/tasks
directory is for unit tests then this feels the most logical place.